button: allow to configure button GPIO in project
Now that multiple variants of ESP32 are supported, it does not make
sense anymore to hardcode the button gpio in the code (the current
hardcoded GPIO isn't even exposed on Xiao ESP32C3).

Set the GPIO number as a project configuration. Set the default value
relevant for Xiao ESP32C3 (button wired on pin D0)

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>