button: replace polling with proper interrupt-driven push detect
The application is currently polling (pretty aggressively) the button
state to detect a button push. Replace the whole logic with a proper
interrupt driven logic.

The button is pretty noisy considering its electrical properties: it
triggers multiple interrupts when pressing it, and it also trigger
interrupts when releasing the button, even if the interrupt is
configured as falling edge. ESP32 chips integrate two types of filters
to help with this, but it appears inneficient in the buzzer case:
- a simple "pin glitch filter" which is not configurable: it will filter
  glitches that are shorter than 2 ticks, and so it does not match the
  noise model for the UBT buzzers
- a "flex glitch filter", which has configurable window and delay.
  Unfortunately this feature is only supported in more featurful
  versions of the chip, and so it is not present in esp32c3 for example

As a consequence, configure a poor man's filter on software side: when
an interrupt is detected, immediately disable interrupts and wake up a
button task which will perform the deboucing. On wake-up:
- wait a small delay to debounce the interrupt
- then read the actual button state. If it is a push, emit the
  notification
- finally, re-enable interrupts

Signed-off-by: Alexis Lothoré <alexis.lothore@gmail.com>